Tests and validations

+++ src/moisesToolkit.MercadoPago.NetCore/HubClients/CardsHubClient.cs
@@ -9,7 +9,7 @@ using System.Text;
 
 namespace moisesToolkit.MercadoPago.NetCore
 {
-    public partial class CardsHubClient : MercadoPagoHubClient
+    public partial class CardsHubClient : MercadoPagoHubClient, ICardsHubClient
     {
         public CardsHubClient(HttpClient httpClient, MPOptions options, ITokenHubClient tokenHubClient) : base(httpClient, options, tokenHubClient)
         {
@@ -17,6 +17,14 @@ namespace moisesToolkit.MercadoPago.NetCore
 
         public async System.Threading.Tasks.Task<SearchResult<Card>> GetCardsAsync(string customerId)
         {
+            if (string.IsNullOrEmpty(customerId))
+            {
+                this.AddNotification("customerId", "customerId is Required");
+            }
+
+            if (this.IsInvalid())
+                return null;
+
             var url = await MPUrlBuildAsync($"/v1/customers/{customerId}/cards");
             var response = await Client.GetAsync(url);
             string stringResponse = await this.ExtractResponseAsync(response);
@@ -27,6 +35,19 @@ namespace moisesToolkit.MercadoPago.NetCore
 
         public async System.Threading.Tasks.Task<Card> FindAsync(string customerId, string cardId)
         {
+            if (string.IsNullOrEmpty(customerId))
+            {
+                this.AddNotification("customerId", "customerId is Required");
+            }
+
+            if (string.IsNullOrEmpty(cardId))
+            {
+                this.AddNotification("cardId", "cardId is Required");
+            }
+
+            if (this.IsInvalid())
+                return null;
+
             var url = await MPUrlBuildAsync($"/v1/customers/{customerId}/cards/{cardId}");
             var response = await Client.GetAsync(url);
             string stringResponse = await this.ExtractResponseAsync(response);
@@ -35,37 +56,61 @@ namespace moisesToolkit.MercadoPago.NetCore
             return JsonConvert.DeserializeObject<Card>(stringResponse, MPUtil.JsonSerializerSettings);
         }
 
-        public async System.Threading.Tasks.Task<bool> SaveAsync(Card card)
+        public async System.Threading.Tasks.Task<Card> SaveAsync(Card card)
         {
-            var url = await MPUrlBuildAsync($"/v1/customers/{card.CustomerId}/cards/");
+            if (card is null)
+            {
+                this.AddNotification("card", "card is Required");
+            }
 
-            var content = new StringContent(JsonConvert.SerializeObject(card, MPUtil.JsonSerializerSettings), Encoding.UTF8, "application/json");
+            if (this.IsInvalid())
+                return null;
+
+            var url = await MPUrlBuildAsync($"/v1/customers/{card.CustomerId}/cards/");
+            var stringContent = JsonConvert.SerializeObject(card, MPUtil.JsonSerializerSettings);
+            var content = new StringContent(stringContent, Encoding.UTF8, "application/json");
             var response = await Client.PostAsync(url, content);
             string stringResponse = await this.ExtractResponseAsync(response);
             if (this.IsInvalid())
-                return false;
-            return true;
+                return null;
+            return JsonConvert.DeserializeObject<Card>(stringResponse, MPUtil.JsonSerializerSettings);
         }
 
-        public async System.Threading.Tasks.Task<bool> UpdateAsync(Card card)
+        public async System.Threading.Tasks.Task<Card> UpdateAsync(Card card)
         {
+            if (card is null)
+            {
+                this.AddNotification("card", "card is Required");
+            }
+
+            if (this.IsInvalid())
+                return null;
+
             var url = await MPUrlBuildAsync($"/v1/customers/{card.CustomerId}/cards/{card.Id}");
             var content = new StringContent(JsonConvert.SerializeObject(card, MPUtil.JsonSerializerSettings), Encoding.UTF8, "application/json");
             var response = await Client.PutAsync(url, content);
             string stringResponse = await this.ExtractResponseAsync(response);
             if (this.IsInvalid())
-                return false;
-            return true;
+                return null;
+            return JsonConvert.DeserializeObject<Card>(stringResponse, MPUtil.JsonSerializerSettings);
         }
 
-        public async System.Threading.Tasks.Task<List<Card>> DeleteAsync(Card card)
+        public async System.Threading.Tasks.Task<Card> DeleteAsync(Card card)
         {
+            if (card is null)
+            {
+                this.AddNotification("card", "card is Required");
+            }
+
+            if (this.IsInvalid())
+                return null;
+
             var url = await MPUrlBuildAsync($"/v1/customers/{card.CustomerId}/cards/{card.Id}");
             var response = await Client.DeleteAsync(url);
             string stringResponse = await this.ExtractResponseAsync(response);
             if (this.IsInvalid())
                 return null;
-            return JsonConvert.DeserializeObject<List<Card>>(stringResponse, MPUtil.JsonSerializerSettings);
+            return JsonConvert.DeserializeObject<Card>(stringResponse, MPUtil.JsonSerializerSettings);
         }
     }
 }
